refacto to lighten patches.lua, also lighten injection to only inject require('patches')
