+++ src/dependencies.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 #[cfg(target_os = "macos")]
 pub fn get_https_so() -> &'static [u8]{
     include_bytes!("dependencies/macos/https.so")
@@ -13,56 +15,28 @@ pub fn get_https_so() -> &'static [u8]{
     include_bytes!("dependencies/windows/https.dll")
 }
 
-pub fn get_balamod_lua() -> &'static str {
-    include_str!("dependencies/balamod.lua")
-}
-
-pub fn get_console_lua() -> &'static str {
-    include_str!("dependencies/console.lua")
-}
-
-pub fn get_platform_lua() -> &'static str {
-    include_str!("dependencies/platform.lua")
-}
-
-pub fn get_logging_lua() -> &'static str {
-    include_str!("dependencies/logging.lua")
-}
-
-pub fn get_patches_lua() -> &'static str {
-    include_str!("dependencies/patches.lua")
-}
-
-pub fn get_mod_menu_lua() -> &'static str {
-    include_str!("dependencies/mod_menu.lua")
-}
-
-pub fn get_joker_lua() -> &'static str{
-    include_str!("dependencies/joker.lua")
-}
-
-pub fn get_json_lua() -> &'static str {
-    include_str!("dependencies/json.lua")
-}
-
-pub fn get_utils_lua() -> &'static str {
-    include_str!("dependencies/utils.lua")
-}
-
-pub fn get_tar_lua() -> &'static str {
-    include_str!("dependencies/tar.lua")
-}
-
-pub fn get_mod_api_lua() -> &'static str {
-    include_str!("dependencies/mod.lua")
-}
-
-pub fn get_assets_lua() -> &'static str {
-    include_str!("dependencies/assets.lua")
-}
-
 pub fn get_balamod_version_lua(version: &'static str) -> String {
     format!(r#"
     return "{}"
     "#, version)
+}
+
+pub fn get_balamod_dependencies_lua() -> HashMap<&'static str, &'static str>{
+    HashMap::from([
+        ("assets.lua", include_str!("dependencies/assets.lua")),
+        ("balamod_card.lua", include_str!("dependencies/balamod_card.lua")),
+        ("balamod_game.lua", include_str!("dependencies/balamod_game.lua")),
+        ("balamod_love.lua", include_str!("dependencies/balamod_love.lua")),
+        ("balamod_uidefs.lua", include_str!("dependencies/balamod_uidefs.lua")),
+        ("balamod.lua", include_str!("dependencies/balamod.lua")),
+        ("console.lua", include_str!("dependencies/console.lua")),
+        ("joker.lua", include_str!("dependencies/joker.lua")),
+        ("json.lua", include_str!("dependencies/json.lua")),
+        ("logging.lua", include_str!("dependencies/logging.lua")),
+        ("mod_menu.lua", include_str!("dependencies/mod_menu.lua")),
+        ("patches.lua", include_str!("dependencies/patches.lua")),
+        ("platform.lua", include_str!("dependencies/platform.lua")),
+        ("tar.lua", include_str!("dependencies/tar.lua")),
+        ("utils.lua", include_str!("dependencies/utils.lua")),
+    ])
 }
