+++ src/dependencies/balamod_uidefs.lua
@@ -1,0 +1,48 @@
+local game_create_UIBox_main_menu_buttons = create_UIBox_main_menu_buttons
+
+function create_UIBox_main_menu_buttons()
+    local t = game_create_UIBox_main_menu_buttons()
+    local modBtn = {
+        n = G.UIT.R,
+        config = {
+            align = "cm",
+            padding = 0.2,
+            r = 0.1,
+            emboss = 0.1,
+            colour = G.C.L_BLACK,
+        },
+        nodes = {
+            {
+                n = G.UIT.R,
+                config = {
+                    align = "cm",
+                    padding = 0.15,
+                    minw = 1,
+                    r = 0.1,
+                    hover = true,
+                    colour = G.C.PURPLE,
+                    button = 'show_mods',
+                    shadow = true,
+                },
+                nodes = {
+                    {
+                        n = G.UIT.T,
+                        config = {
+                            text = "MODS",
+                            scale = 0.6,
+                            colour = G.C.UI.TEXT_LIGHT,
+                            shadow = true,
+                        },
+                    },
+                },
+            },
+        },
+    }
+
+    local insertIndex = #t.nodes[2].nodes
+    if not G.F_ENGLISH_ONLY then
+        insertIndex = insertIndex - 1
+    end
+    table.insert(t.nodes[2].nodes, insertIndex, modBtn)
+    return t
+end
